fix underline for sdknav
