+++ src/components/SdkNav/styles.module.css
@@ -1,68 +1,3 @@
-.sdkNav {
-  --sdk-nav-accent: var(--interactive-primary-surface, #444ce7);
-  --sdk-nav-gradient: linear-gradient(255.4deg, #444ce7 0%, #b664ff 100%);
-
-  position: sticky;
-  top: var(--navbar-primary-height);
-  z-index: 150;
-  height: var(--secondary-nav-height);
-  background: var(--ifm-navbar-background-color);
-  border-bottom: 1px solid var(--ifm-toc-border-color);
-  font-size: 0.875rem;
-}
-
-.inner {
-  display: flex;
-  align-items: center;
-  height: 100%;
-  padding: 0 var(--ifm-navbar-padding-horizontal);
-  overflow-x: auto;
-  scrollbar-width: none;
-}
-
-.inner::-webkit-scrollbar {
-  display: none;
-}
-
-/* Keep the two-background ring: a masked pseudo-element fills solid in
- * production, autoprefixer drops mask-composite. */
-.group {
-  display: inline-flex;
-  gap: 0.375rem;
-  padding: 0.3125rem;
-  border: 1px solid transparent;
-  border-radius: 999px;
-  background:
-    linear-gradient(var(--ifm-navbar-background-color), var(--ifm-navbar-background-color)) padding-box,
-    var(--sdk-nav-gradient) border-box;
-}
-
-.sdk {
-  display: inline-flex;
-  align-items: center;
-  gap: 0.5rem;
-  padding: 0.4375rem 1rem;
-  white-space: nowrap;
-  font-weight: 500;
-  color: var(--ifm-navbar-link-color);
-  text-decoration: none;
-  border-radius: 999px;
-  transition:
-    color 120ms ease,
-    background-color 120ms ease;
-}
-
-.sdk:hover {
-  color: var(--ifm-navbar-link-hover-color);
-  text-decoration: none;
-}
-
-.sdkActive,
-.sdkActive:hover {
-  color: #fff;
-  background: var(--sdk-nav-accent);
-}
-
 .icon {
   width: 1.125rem;
   height: 1.125rem;
